% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{neo4j_conn_status}
\alias{neo4j_conn_status}
\title{Check if Neo4j Database is Running}
\usage{
neo4j_conn_status(conn)
}
\description{
Check if Neo4j Database is Running
}
\seealso{
\code{\link[neo4jshell]{neo4j_status}}
}
