% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{stop_neo4j}
\alias{stop_neo4j}
\title{Stop Neo4j Database}
\usage{
stop_neo4j(
  db_key,
  neo4j_home = "~/Library/Application Support/com.Neo4j.Relate/Data/dbmss"
)
}
\description{
Stop Neo4j Database
}
\seealso{
\code{\link[neo4jshell]{neo4j_stop}}
}
