% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{store_db_name}
\alias{store_db_name}
\title{Store Database Name}
\usage{
store_db_name(
  new_db_name,
  new_db_key,
  neo4j_home = "~/Library/Application Support/com.Neo4j.Relate/Data/dbmss"
)
}
\description{
Store Database Name
}
\seealso{
\code{\link[R.cache]{findCache}},\code{\link[R.cache]{saveCache}},\code{\link[R.cache]{loadCache}}
\code{\link[tibble]{c("tibble", "tibble")}},\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{coalesce}}
\code{\link[stringr]{str_remove}}
}
