% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{start_neo4j}
\alias{start_neo4j}
\title{Start Neo4j Database}
\usage{
start_neo4j(
  db_key = "dbms-3e732dbc-5fd2-4f54-b6d5-dfaca7494d7c",
  db_name,
  neo4j_home = "~/Library/Application Support/com.Neo4j.Relate/Data/dbmss",
  uid = "neo4j",
  pwd = "admin",
  verbose = TRUE
)
}
\description{
Start Neo4j Database
}
\seealso{
\code{\link[readr]{read_lines}}
\code{\link[neo4jshell]{neo4j_start}}
\code{\link[glue]{glue}}
\code{\link[stringr]{str_replace}}
}
